Fix issue with PATH causing test failures
I was seeing an issue where modifying PATH caused the Ruby plugin
tests to fail. I fixed that and while digging around fixed an issue
that would cause the tests to behave weirdly if run from any other
location.

This solves one of the issues noted in #687, but I haven't tried
running the new tests from that PR with these changes.